Move ClusterStateBundle to vdslib, allowing it to be used outside storage module.
