+++ storage/src/tests/common/testnodestateupdater.cpp
@@ -1,7 +1,7 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "testnodestateupdater.h"
-#include <vespa/storage/common/cluster_state_bundle.h>
+#include <vespa/vdslib/state/cluster_state_bundle.h>
 
 namespace storage {
 
@@ -14,7 +14,7 @@ TestNodeStateUpdater::TestNodeStateUpdater(const lib::NodeType& type)
 
 TestNodeStateUpdater::~TestNodeStateUpdater() = default;
 
-std::shared_ptr<const ClusterStateBundle>
+std::shared_ptr<const lib::ClusterStateBundle>
 TestNodeStateUpdater::getClusterStateBundle() const
 {
     return _clusterStateBundle;
@@ -23,7 +23,7 @@ TestNodeStateUpdater::getClusterStateBundle() const
 void
 TestNodeStateUpdater::setClusterState(lib::ClusterState::CSP c)
 {
-    _clusterStateBundle = std::make_shared<const ClusterStateBundle>(*c);
+    _clusterStateBundle = std::make_shared<const lib::ClusterStateBundle>(*c);
     for (uint32_t i = 0; i < _listeners.size(); ++i) {
         _listeners[i]->handleNewState();
     }
