+++ storage/src/vespa/storage/common/CMakeLists.txt
@@ -3,7 +3,6 @@ vespa_add_library(storage_common OBJECT
     SOURCES
     bucketmessages.cpp
     bucketoperationlogger.cpp
-    cluster_state_bundle.cpp
     content_bucket_space.cpp
     content_bucket_space_repo.cpp
     distributorcomponent.cpp
