+++ vdslib/src/vespa/vdslib/state/CMakeLists.txt
@@ -7,5 +7,6 @@ vespa_add_library(vdslib_state OBJECT
     diskstate.cpp
     nodestate.cpp
     clusterstate.cpp
+    cluster_state_bundle.cpp
     DEPENDS
 )
